Move migration logic to a separate class.
