+++ library/src/main/java/com/novoda/downloadmanager/Migrator.java
@@ -1,0 +1,9 @@
+package com.novoda.downloadmanager;
+
+public interface Migrator {
+    void migrate();
+
+    interface Callback {
+        void onMigrationComplete();
+    }
+}
